refactor after translation implementation
